fixed problem in AddonManager which caused lint to fail, refractored AddonManager and separated tests for array, math and statistical addon in three files
