+++ src/addons/index.ts
@@ -1,5 +1,7 @@
+import { AddonManager as Addonable } from './AddonManager';
 import { MathAddon as MathAdd } from './MathAddon';
 import { StatisticalAddon as StatisticalAdd } from './StatisticalAddon';
 
 export const MathAddon = MathAdd;
 export const StatisticalAddon = StatisticalAdd;
+export const AddonManager = Addonable;
